requirements.txt
streamlit
tensorflow
matplotlib
seaborn
scikit-learn
mlflow
dagshub
pillow
markdown2
pdfkit
transformers
torch